I hate Discord?
Discord apparently does not allow both embeds and attachment embeds to be displayed in a single message, so a choice to be made is whether to display the user's message or the links possibly sent with it.

I've opted for the user message.

A work-around is to send two starboard messages, but that would break the current design principle of 'one message; one starboard message' and would require (likely) an extensive overhaul of the current code that could take me a few hours.

Perhaps it's better left when I rewrite this God-forsaken codebase in an actually good language, like C#.